Update: Shared 페이지 ts로 변경
