+++ src/pages/Shared.tsx
@@ -1,0 +1,69 @@
+import React from "react";
+import ConHeader from "../components/ConHeader/ConHeader";
+import Search from "../components/Search/Search";
+import CardList from "../components/CardList/CardList";
+import { useEffect, useState } from "react";
+import { fetchData } from "../fetchUtils";
+
+interface FolderData {
+  id: number;
+  name: string;
+  owner: {
+    id: number;
+    name: string;
+    profileImageSource: string;
+  };
+  links: {
+    id: number;
+    createdAt: string;
+    url: string;
+    title: string;
+    description: string;
+    imageSource?: string;
+  }[];
+  count: number;
+}
+
+interface CardListData {
+  id: number;
+  createdAt: string;
+  url: string;
+  title: string;
+  description: string;
+  imageSource: string;
+}
+
+interface FetchDataResponse {
+  cardListData: CardListData[];
+  folderData: FolderData[];
+}
+
+function Shared() {
+  const [folderData, setFolderData] = useState<FolderData[]>([]);
+  const [cardListData, setCardListData] = useState<CardListData[]>([]);
+
+  console.log(cardListData);
+  useEffect(() => {
+    async function fetchDataAndSetState() {
+      const response: FetchDataResponse | undefined = await fetchData();
+      if (response) {
+        const { cardListData, folderData } = response;
+        setFolderData(folderData);
+        setCardListData(cardListData);
+      }
+    }
+    fetchDataAndSetState();
+  }, []);
+
+  return (
+    <div className="content-wrap">
+      <ConHeader folderData={folderData} />
+      <div className="wrap">
+        <Search />
+        <CardList cardListData={cardListData} />
+      </div>
+    </div>
+  );
+}
+
+export default Shared;
